Explicit cast to short
- Handles some weirdness where NMS can't seem decide return values for
mapviews should be short or int.

+++ src/main/java/me/Fupery/ArtMap/IO/CompressedMap.java
@@ -18,7 +18,7 @@ public class CompressedMap extends MapId {
     }
 
     public static CompressedMap compress(MapView mapView) {
-        return compress(mapView.getId(), Reflection.getMap(mapView));
+		return compress((short) mapView.getId(), Reflection.getMap(mapView));
     }
 
     public static CompressedMap compress(short mapId, byte[] map) {
